# src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: phenix API
  version: 0.0.0
  title: phenix
  termsOfService: http://swagger.io/terms/
  contact:
    url: https://github.com/sandialabs/sceptre-phenix
  license:
    name: GNU General Public License v3.0
...
paths:
...
  "/disks/inject":
    post:
      tags:
        - Disks
      summary: Inject miniccc or minirouter into a disk image
      operationId: injectMiniExe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - disk
              properties:
                exe:
                  type: string
                  default: /phenix/miniccc
                disk:
                  type: string
                  description: Absolute path or a filename relative to /phenix/vmdb
                svc:
                  type: string
      responses:
        "204":
          description: executable injected
        "400":
          description: invalid request
        "403":
          description: forbidden
  "/images/{name}/build":
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Disks
      summary: Get the status of the most recent image build
      operationId: getImageBuild
      responses:
        "200":
          description: image build status
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    enum:
                      - building
                      - complete
                      - failed
                  error:
                    type: string
        "403":
          description: forbidden
        "404":
          description: no build found
    post:
      tags:
        - Disks
      summary: Build a VM disk image
      operationId: buildImage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                verbosity:
                  type: integer
                cache:
                  type: boolean
                dry_run:
                  type: boolean
                output:
                  type: string
                  default: /phenix/vmdb
      responses:
        "202":
          description: build accepted; poll the URL in the Location header for status
          headers:
            Location:
              schema:
                type: string
        "400":
          description: invalid request
        "403":
          description: forbidden
        "409":
          description: build already in progress
...
servers:
  - url: http://localhost:3000/api/v1
...